main/requirements.txt
bs4==0.0.1
cryptography==3.3.2
Django==4.2.2
gensim==3.8.3
google-cloud-vision==3.4.4
openpyxl==3.1.2
Pillow==9.5.0
screeninfo==0.8.1
selenium==4.10.0